PF-2451: Address some of the sonarqube scan reliability issues (#1042)
* add build task and update gha

* gradle

* address security concern

* another security hotspot

* lint

+++ service/src/main/java/bio/terra/workspace/common/utils/RetryUtils.java
@@ -64,7 +64,7 @@ public class RetryUtils {
             sleepDuration.toSeconds(),
             endTime);
         TimeUnit.MILLISECONDS.sleep(sleepDuration.toMillis());
-        long increaseMillis = Double.valueOf(factorIncrease * sleepDuration.toMillis()).longValue();
+        long increaseMillis = (long) (factorIncrease * sleepDuration.toMillis());
         sleepDuration = sleepDuration.plusMillis(increaseMillis);
         if (sleepDuration.compareTo(sleepDurationMax) > 0) {
           sleepDuration = sleepDurationMax;
@@ -133,7 +133,7 @@ public class RetryUtils {
           throw new Exception();
         }
         TimeUnit.MILLISECONDS.sleep(sleepDuration.toMillis());
-        long increaseMillis = Double.valueOf(factorIncrease * sleepDuration.toMillis()).longValue();
+        long increaseMillis = (long) (factorIncrease * sleepDuration.toMillis());
         sleepDuration = sleepDuration.plusMillis(increaseMillis);
         if (sleepDuration.compareTo(sleepDurationMax) > 0) {
           sleepDuration = sleepDurationMax;
